[doc] TODOs removed
As per the list below, they are all checked or we changed the way we
want to do things.

- [x] Quotes in site name / tagline, e.g. `Nicolas's site`, are not 
escaped properly.
- [x] We need to have a way to make the operator manage Ingresses / 
Deployment to balance the loads / the numbers of site per pods
- [x] As for the Ingresses and Deployments, same have to happen with 
MariaDB (the database server)
- [~] We have to manage the languages (which could be changed by the 
users, so this imply some knowledges of the modification made at user 
level, see next point)
- [~] We need to have some sort of cron task that read the actual 
configuration, i.e. language, that are active on a site and write thoses 
in the kubernetes object (status / annotation)
- [x] The operator has to create the Databases backup/restore
- [x] The life cycle of a WordPressSite has to be fully managed by the 
operator: introduce a thombstone/archive status that can be revived, 
duplicate a site
- [x] The CRD should have more filter field (the one that are shown on 
the `kubectl get wp` output), such as `path`, etc
- [-] wp-veritas (a clickable interface to manage WordPress sites)