[feature] Recreate CRD when deleted
As per
https://github.com/nolar/kopf/issues/702#issuecomment-2511251889 we
need to be extra careful in `@on.delete()` callbacks. To wit:

- it appears that the `optional=True` parameter to `on.delete()` does
exactly what it says on the tin → no event is received 😁
- a specific filter must be set using `field` and `value`, lest Kopf
scribble onto each and every CRD in the cluster (although a little
less scribbling happens when `optional=True`, this behavior cannot
easily be disabled completely)
- the `on.delete()` hook must complete before we attempt to recreate
the object, or deadlock occurs.