[fix] Select resources tightly
We don't want to watch *all* `RoleBinding`s, *all* `Deployment`s etc.
etc. — even if we promptly `return` out of those `@kopf.daemon`
handlers that we aren't interested in. For starters, this pollutes the
logs; and also, this breaks multi-namespace support because it seems
that Kopf doesn't stack the `@kopf.daemon`s that have the exact same
selector!