[fix] “second-round bug” for namespaced objects
... whence everything worked juuust right the first time one created
the first `WordpressSite` in an empty namespace (or one started the
operator with a namespace already populated with one or more
`WordpressSite`s); then when this or these `WordpressSite`(s) would be
deleted, the controller would clean up the namespace as desired; and
then when one recreated a `WordpressSite` in the empty namespace, the
`ExistenceReconciler` would stop doing the second half its job i.e. it
would create the object, but fail to recreate it when deleted.

This one was... a bit of a conundrum to crack. Thanks to ChatGPT, I
was able to confirm my intuition that the second `ExistenceReconciler`
instance would fail to effectively start its `@kopf_daemon`, because
the former one which shares the same characteristics (same
`api_version`, `kind` and `when` function) still lingers around
for (worst-case) another 3600 seconds.

- Refactor `KubernetesObjectData().kopf_daemon` to accept extra
kwargs (in fact, the “plain decorator with no parentheses” convention
is no longer available)
- Forward the kwargs to the underlying `@kopf.daemon` call
- Use that to pass a unique `id=` based upon a monotonically increasing
sequence, so that multiple instances of `ExistenceReconciler`
targeting the same object may do their daemon thing at the same time
- Throw in some `cls.__name__` into the ID, so that if another
instance of myself has the same idea later in the code, the IDs won't
clash 😜